tutorial_basics/requirements.txt
python-dotenv>=0.19.0
pydantic-ai>=0.0.21 